6th sem projec/SmartAttendanceAI/requirements.txt
# Web Framework
Flask==3.0.3
Werkzeug==3.0.3

# Database
mysql-connector-python==8.4.0

# Computer Vision & AI
opencv-python==4.10.0.84
face-recognition==1.3.0
dlib==19.24.4
numpy==1.26.4
Pillow==10.4.0

# Deep Learning (Mask Model)
tensorflow==2.17.0

# Analytics & Export
pandas==2.2.2
matplotlib==3.9.0
openpyxl==3.1.5

# Utilities
requests==2.32.3
python-dotenv==1.0.1